replace main.py for better diff
